Subsonic: Fix notice "Only variables should be passed by reference"
This kind of notice could get logged while browsing tracks via Subsonic
API, depending on the PHP version and/or settings. This was because the
PHP built-in function end() was being used incorrectly. According to
the documentation, only "real" arrays should be fed to it and not the
temporary variable created  when an array-returning function is inlined
as parameter of the end() function.

The call still worked as expected, but just emitted this notice.