Refactoring: Move file/memory size formatting from the dedicated class to Util
There used to be a dedicated FileSize class used to convert file sizes (or
other memory sizes) to human-readable format. This was quite an overkill, a
the whole logic is just 3 lines of code. Now this class has been replaced
with a static utility function Util::formatFileSize, which is also more
straight forward to use than the previous construct.