Subsonic: Prevent getting folder info from outside of user music path
Previously, it was possible to obtain subfolders of any folder via
Subsonic API, provided that the caller knew/guessed/brute-forced the
folder IDs. Now the access is limited to folders within the configured
user music path. The breach was not very serious, though, because it
did not enable accessing names or contents of any files, only folders.

refs #725