Subsonic: Don't include "parent" attribute on artist/top-folder entities
This is what the "real" Subsonic server does, as well. Having the
inappropriate parent attribute caused the Subsonic Android app to
misbehave when pressing the arrow icon on the upper left corner of the
screen while viewing the contents of an artist or a top-level folder:
An error message was shown and the UI did not navigate back to the
"index" page.

refs #725