Subsonic: Fix some issues introduced by the commit c98fe53e6
The commit c98fe53e6 "Subsonic: Add the API method getLyrics" caused
some unintended side effects, when it slightly changed the formatting
of XML in XMLResponse class. Empty XML elements were being created
where they were not supposed to be, like in getVideos and getStarred
methods, as well as when fetching the contents of an empty folder.
This, in turn, caused crashes both on Subsonic Android app and DSub.

Hence, the problematic change to XMLResponse was reverted, and the
intentional empty XML element needed in getLyrics is now handled a bit
differently. This new method is also better when using the JSON
formatted results, as the empty element is now shown as "{}" instead of
"[]". This matches the behavior of the "real" Subsonic server.