Lazy-load album cover images
Do not load all the cover images immediately upon loading the Music app
main view. Instead, load only the visible covers and those which are
close to become visible with a bit of scrolling. This should help on
the excessive number of HTTP requests firing when opening a library
with huge number of albums having cover images.

The mechanism relies upon IntersectionObserver and WeakMap classes
which may not be available on older browsers. On such browsers, the
lazy-loading is disabled and everything works as before.

refs #719, #653