Lazy-load track list contents
Track list directive is now populated with the actual contents only when it is
in the viewport or close to it. Registering all the listeners is postponed to
the same time. Furthermore, the track list contents and listeners are cleaned
up when the list element leaves the viewport.

This improves the performance and memory footprint on huge collections while
not interfering too much with the look and feel of the scrolling. This is
enough to make collections of tens of thousands of tracks somewhat usable also
on Chrome, while such collections could previously be used only on Firefox.
Also, on Firefox, it was measured that navigating from Albums view to Folders
view with 9347 folders now took ~35s while it previously took ~65s. Memory
consumption of the folders view with the same test content dropped from ~700
MB to ~300 MB.

Supporting the lazy-loading requires a sufficiently modern browser. On any
version of IE and on other ancient browsers, the lazy-loading is disabled and
everything should work as before.

refs #728