Fix a dead-lock on Folders view deactivation
It was possible to end up in a dead-lock situation if the user navigated to
the Folders view, and then to another view before the Folders view was fully
loaded. In that case, the incremental showing and hiding of the view ran
simultaneously and neither could ever finish. More than 50 folders were
needed to reproduce this.