Substitute <track-list> with <ul> already when compiling the directive
There's now `compile` function defined for the track-list directive. This
hook is run already when angularjs traverses the template, as opposed to the
`link` function which is run separately for each instance created by
ng-repeat.

The `link` function is now defined with key `post` in an object returned by
`compile`. (The normal angularjs `link` function is actually post-link
function and it would be possible to define also pre-link function.)