Move isElementInViewPort from OC_Music_Utils to a new angularjs service
This function didn't really belong to the shared utils, because it
wouldn't have worked correctly, if it was used from the embedded Files
player.

Furthermore, having it in its own service lets us cache the header and
and appView heights. At least on Edge, the profiling showed that
reading these properties from DOM took quite some time. However, it
might be that this just moves the load elsewhere, because it's quite
likely that the browser has some kind of caching logic, as well.