Fix a layout issue with extra wide screens on NC14+
Starting from the Nextcloud version 14, the width of the sidebar has
been limited to 500px at maximum, although normally it's 27% of the
screen width. This limit comes into effect when the screen is extra
wide or the page is zoomed out. We were always reserving 27% of the
screen width for the sidebar, and this resulted in a weird margin
appearing between the sidebar and the main content.