Ignore any events in inViewObserver during page loading
Loads of unnecessary processing happened at least when activating the
Folders view with a huge amount of folders: The first 100 folders were
first added with 0-size, and all of them were evaluated to be within
the view-port. Then they got their actual size after which most of them
left the view-port and the previously built contents were replaced with
placeholders.

refs #728