Optimize view deactivation in Folders and Albums views
There's no need to manually remove the HTML elements within the
trackList and listHeading directives when the scope is destroyed.
Omitting these also makes the view deactivation so much faster, that the
asynchronous "showLess" construct could be removed from the Albums and
Folders views. This improved the view deactivation speed even further.

refs #728