Do not calculate trackList placeholder height from actual child elements
This was a comparatively expensive operation, degrading the scroll
performance on huge collections.

Now the height is always estimated solely based on the number of the
tracks. The style definitions were slightly tweaked to make this
easier, and now the estimates are actually more accurate than the
previous "measured values" (!).