Delay "leaveView" notifications until when scrolling stops
The inViewObserver now delays "leaveView" notifications to its
listeners until when the scrolling stops. This is because replacing
trackLists and listHeadings with the placeholders during scrolling
interfered with the smooth scrolling on huge collections. On the other
hand, there's no hurry in doing the placeholder substitution, so we can
well wait until the scrolling ends.

refs #728