Performance improvement for searching within a playlist
Search only among the tracks belonging to the current playlist when
search conducted in a playlist view. This has a tremendous impact on
the performance if the library is huge but the playlist is small to
medium sized (i.e. does not contain close to as many tracks as the
library as whole).