Fix the the auto-scroll when Music launched from Files
Scrolling to the selected track/album/artist didn't work when the Music
app was launched by selecting a search result in the Files app. The
same problem was present when the Music app was launched from the
embedded music player in the Files app. This has probably got broken
when the lazy-loading/InViewObserver stuff was implemented.