Fix invisible #searchresults container screwing up the mobile layout
In the mobile layout, the #searchresults container had negative impact
on the placement of other elements when a search was executed in the
Music app. In this case, the core gives it "display: block" style, and
it affects the placement of other elements even though it's always
hidden by us. Using "position: absolute" prevents it affecting any
other elements.