Refactor and fix inViewObserver
The previous implementation had such a problem, that it could trigger
the "enterView" event for all instances when navigating from another
view to the Albums view. This, in turn, could cause all the album art
to be loaded at once.

The logic was now changed so that there are two distinct modes: The
range of visible instances is tracked only in the "normal" mode. While
the search is active, and some (or all) of the instances may be hidden,
then the module doesn't even try to track the range of visible
instances. This is because the range may have "holes" in form of hidden
instances, and following this would be too complicated.