Do not shrink down album covers quite so aggressively
The maximum size for the album covers in the Albums view is now 380x380
pixels instead of the previous 200x200 pixels. The previously used
resolution looked a bit rough, especially if the browser or OS was
configured to scale up the UI. The download size of the cover images is
more than doubled, but it is still only some tens of kilobytes per
image. This shouldn't be too much now as we load the cover images on
demand.

refs  #734