Scale cover images with imagecopyresampled instead of imagescale
The PHP function imagecopyresampled seems to create significantly
better looking downscaling results than the other PHP function
imagescale. The produced filesize is the same or even a bit less than
with imagescale.

Furthermore, the cover images are now cropped to square shape on the
back-end, while this was previously done on the front-end. This could
potentially cause transfer and caching some unnecessary data, in case
the image aspect ratio was far from square shape. Also, with
Ampache/Subsonic APIs, it was totally up to the client app to do
reasonable handling of the non-square cover images, and some of them
were not doing all that great job.

#734