Subsonic: Support specifying target size on getCoverArt
Previously, we were always returning the cover art in the default size,
regardless of any size specified by the client. Now the default size is
used only if the client doesn't specify any size. Using the cover cache
is skipped, if the cover art is requested in a specific size.

This change is significant at least for the Subsonic Music Streamer app
on Android, because it requests cover images at 128px size on its album
list view, and assumes to get the images in this size. Previously, as
we were actually returning larger images, the app also drew them in
larger size, which made the list UI look a bit off; especially after we
increased the default size to 380px. What made this worse, was that if
some albums had cover smaller than 380px, then heights of the list rows
varied.