Allow defining the cover image size in config.php
It is now possible to use the key 'music.cover_size' in the config.php file
to define the size to be used for the album cover images. The default size
380px is used if the key is not present in the config file.

Like before, this is jut the maximum size for the cover images to be used in
the Albums view of the web UI. Images smaller than the given size are not
scaled up, but images larger are scaled down. The same image size is used in
the Ampache and Subsonic APIs, unless the client specifies another size.

refs #734