Check for unscanned files on web UI only after loading the collection
Previously, checking the availability of any unscanned music files happened
in parallel with preparing the collection.json. Both of these operations are
CPU intensive and they slowed down each other.

Now the collection is loaded first and the unscanned files are checked
afterwards. On huge collection, this makes the collection appear on the
screen a bit sooner, and reduces the risk of HTTP timeout breaking either of
the operations. On the downside, it now takes more time from the start of the
page load until the scanning of the new files can start (provided that there
are already lot of scanned files).

On small to medium sized collections, this makes no practical difference.
Both the positive and negative effects concern only the huge collections.