Match each word of multi-part search query separately
In case the search query used within the Music app contains many whitespace
separated words, each word is now matched against the entities separately.
The search matches the entity, if each part can be found.

The user may bail out of this by surrounding the search query with quotation
marks ("). In this case, the query is treated as a single string which must
be found as such, including the contained whitespace characters.