Refactor libraryService to enable more efficient access to parent entities
When the collection is fetched, the libraryService now sets up parent
references to each entity so that it is easy to navigate from track to album
and from album to artist.

Similarly, when foldres are fetched, the parent references are set to allow
navigating from a track to its containing folder.