Support combining parts of artist, album, and track details in search query
A multi-word search query can now be matched so that e.g. one word matches
the artist name, another matches the album name, third matches album year,
and fourth matches track title. That is, the user can use any combination of
details from a track (or album) and their parent entities to define the items
they want to see.

This required rather extensive changes on the search logic. Technically, we
are no longer searching separately albums and artists. Instead, we are only
searching tracks but consider the parent fields, too, when checking the
match. One side benefit of this  change was, that the maximum number of
matches to show could be increased, because we no longer use the costly union
operation.