Limit the max number of matches on the back-end search provider
The architecture of the ownCloud search providers is not really designed for
a huge number of matches. The UI loads at most 30 matches at time, but all
the matches are queried from the Provider class on each 30 item page. The UI
is so clumsy, that the user probably wouldn't care to load more than the
first 30 matches anyway. Hence, the Provider would do a lot of unnecessary
work if there were e.g. thousands of matches returned to the core.

The back-end search provider does the search even when we are using the Music
app UI, where the results from the back-end are disregarded in favor of the
filtering made on the front-end. Hence, the Provider could have been
causing lots of totally unnecessary CPU load on the server.