Fix a background color problem on NC18 with dark theme
The Nextcloud 18 server includes a style definition for the body element
which really seems like a bug: there are two definitions for the property
"background-color" on the same css selector, and hence, the second one
overrides the first one. The second definition is some "linear-gradient"
which doesn't seem to be a supported value for the "background-color"
property.

Fixes by overriding the background-color definition from the server with a
definition of our own, which in turn is a copy of the shadowed definition
made by the server.

Note that the css variable "--color-main-background" used is not defined on
ownCloud and older versions of Nextcloud. To minimize any side effects caused
by this, we apply our rule only on body element having the class "dark". This
class is added to the body only by NC18 in the dark mode.

refs #739