Fix 'All tracks' view not always showing the tracks
In case there were so few tracks that no scrollbar appeared to the 'All
tracks' view, it could happen that the tracks were left invisible after
activating the view. This was because no 'onScroll' event fired *after* the
ng-repeated tracklists had been laid on the screen.

There were two distinct cases where this could happen:
a) when reloading the browser page in the 'All tracks' view
b) when navigating from another view to the 'All tracks' view

In scenario a), the problem occurred every time, but in the scenario b),
there was some random factor.