Improve the ordering of tags in the details pane
Many of the metadata fields have multiple possible names, depending at
least on the tag format ID3v1/ID3v2/APE, and possibly also on the
application used to write the tags. Attempt to consider all the
possible tag names to show the tags in a logical and consistent order in
the details pane.