Do not treat each disc of a multi-disc album as a separate album
Previously, each disc of a multi-disc album was indexed and shown as if
it was its own album. Each Album entity in the database carried a disc
number, and the entity could have exactly one of those.

Now this was changed so that the discs are not stored as separate Album
entities. Instead, tracks of each disc are stored in the same Album
entity, and each Track entity carries also the disc number.

On the web UI, the multi-disc albums are now shown as one album, and
the disc number is shown in front of the track number. E.g.
1-1. First track name
1-2. Second track name
1-3. Third track name
2-1. First track of the second disc
2-2. Second track of the second disc
...
One single-disc albums, the disc number is not visible on the UI.

In the Ampache, Subsonic, and Shiva APIs, the disc numbers are not
used. The Ampache and Subsonic APIs would offer some mechanisms to
handle the multi-disc albums, but because of the lack of the client
support, there's no point in trying to use those. Instead, we fabricate
the track number given in the APIs so that it includes also the disc
number information: the number used in the API is
(100 * disc_number + track_number).
The fabricated number is used only for multi-disc albums.

On software update, a migration script is run. It moves the disc
numbers from the Album entities to the Track entities, and combines the
Album entities of the multi-disc Albums.

refs #680