Fix the SQL query for getting album disc count
The query actually didn't make much sense in case it was limited to
given albumIds. However, it happened to work on SQLite and also on MySQL
when it was set up with more relaxed syntax rules. Adding the option
`sql_mode = 'ONLY_FULL_GROUP_BY'` made MySQL reject the query.

This error was introduced in the commit 4a468d327.