Fix an update breaking bug in disk number migration
There was a logic flaw which could break the software update process on
certain libraries. The problem could occur, if the library had
multi-disc albums, where the casing of the album name was different
between discs. E.g. if disc 1 was named "Some Album" and disc 2 was
named "SOME ALBUM", then they may not get correctly merged, and
consequently, the unique constraint of the DB got violated when
recalculating the hashes.

Note that this doesn't solve the whole issue for all systems. This is
because the case insensitive comparison doesn't work properly on SQLite
in case the name contains non-ASCII characters.

refs #748