Catch any unexpected exceptions in the disk number migration
Add one more safeguard to the disk number migration script, and catch
any unexpected exceptions on the main level of the script.

The downside of this is, that some problems may go unnoticed, and
unexpected behavior is observed later when using the Music app. But at
least the script won't now block the installation process, which is
*bad* because there's no easy way to rollback to the previous app
version. And even disabling an app with update problems requires using
the command line.

refs #748