Fix notice about invalid array access on track details
In case the track in question has no embedded cover art, the
DetailsHelper class used to do invalid array access when the user
opened the details pane. On older PHP versions, this silently returned
null, but on PHP 7.4, there is a notice emitted. This appeared to the
log (at least on Nextcloud 18) with level 'error'.