Subsonic: Support starring tracks, albums, and artists
Added support for Subsonic API methods `star`, `unstar`, `getStarred`,
and `getStarred2`. In addition, the API methods `getAlbumList` and
`getAlbumList2` now support the argument value `type=starred`.

The starring information is not shown on the web UI for the time being,
nor is the starring/unstarring possible via the web UI.

refs #750