Remove unused function ArtistMapper::findByName()
This function was not used and not even revealed on the BusinessLayer.
Furthermore, the function was bad by design, as it was intended to be
used to find exactly one match, but supported also "fuzzy mode" where
there may easily be multiple matches. Hence, it's difficult to imagine
any real-life use for such function.