Refactor and fix Mappers
BaseMapper now defines protected function `selectUserEntities` which
replaces the function `makeSelectQuery` previously defined in each and
every derived class. It is built upon another new function
`selectEntities`, which may be overridden by the derived class when
necessary (currently done only by TrackMapper).

In addition to removing redundancy, this fixes the issue that the
function BaseMapper::findAllStarred didn't work quite correctly on
TrackMapper: the result objects didn't contain the `filename` and
`size` fields. Now this gets taken care of as the BaseMapper uses the
version of `selectEntities` overridden in TrackMapper.