Subsonic: Fix/tweak the lyrics fetching logic
The Subsonic `getLyrics` API method is silly in how it identifies
tracks by the combination of title and artist, and not by the track ID.
There is no logical reason, why the combination of title and artist
would always uniquely identify a track: consider e.g. the same track
being present on an album, a single, and a compilation.

Previously, the Subsonic `getLyrics` returned an empty result if there
were more than one match for the supplied title/artist pair. Now it
returns lyrics extracted from the first matching track. This is still
not optimal, because it is possible that there are two matching tracks,
but only second of them would contain any lyrics.