Integrate the web UI with the (Chrome) media control panel
Integrate the Music web UI with the mediaSession API. At the moment,
Chrome has support for this, and Firefox should have support starting
from version 76 (the current version is 75, and there's a partial
support available if the advanced configuration switches
dom.media.mediasession.enabled and media.hardwaremediakeys.enabled are
set to TRUE).

With this, desktop Chrome shows a media control button, which pops up a
media control panel with play/pause/next/prev/forward/backward buttons.
The panel also shows the track metadata. Chrome on Android shows the
corresponding controls and info in the status "curtain". Both Windows
10 and Android show the media controls also on the lock screen.

Furthermore, this enables using the special multimedia keys
play/pause/etc on the computer keyboard if such are available. The
browser doesn't even need to have focus for these to work.