Drop the SoundManager2 library in favor of using HTML5 API directly
We were using the SoundManager2 library in the "html5 only" mode. As
such, it was dubious at best, if the library brought any added value.
The HTML5 audio API is simple enough to use for basic functionality,
and dropping the SM2 library actually allowed us to simplify the
PlayerWrapper code on some aspects (e.g. asynchronous init function is
no longer needed).