Detach from mediaSession API (on Chrome) when playing ends
When the player is stopped (e.g. when the last track of the queue has
been played or when the user presses X button on the embedded player),
the src attribute of the HTML audio element is now cleared. This isn't
quite as simple as one would think, which is described in the comments
in more detail.

Clearing the audio src has the effect of detaching the player form the
Chrome audio controls. Hence, Chrome no longer captures the dedicated
playback buttons of the keyboard or show the controls on the OS lock
screen after playing has been stopped.