Pause ongoing playback immediately when next/prev pressed
This got semi-accidentally changed with the commit 6eece2d29a so that
the current playback went on until the newly selected track actually
started playing. This could be indefinitely long time, if the user just
kept tapping the next/prev button. Also, the progress bar wasn't
properly cleared in this case.