Fix playback state clearing when 'stop' received from mediaSession
When the playback was stopped via the mediaSession (e.g. hardware
'stop' button on the keyboard), the playlistService state was not
properly cleared. Hence, the UI was left into inconsistent state, where
the pause icon was still being shown in front of the previously played
track and the previosly playing view in the navigation pane.