Major refactoring: Add DB table *PREFIX*music_genres
Make genre a "first class citizen" in the system with own DB table,
Entity class, Mapper, and BusinessLayer. The main benefit of this is
that we get persistent numeric IDs for the genres. This will come handy
when we add the genre support to the Ampache API. Secondary benefit
is, that we can classify the genres in case-insensitive manner, but
still have user-specific casing for the genre names.