Fix unable to open details pane if getID3 can't extract any metadata
This was detected on one Opus Vorbis file where getID3 emitted an error
message "Unknown opus version number (only accepting 1-15)". In that
case, the details pane used to be left in the "loading" state
indefinitely. Now the pane opens, and shows the technical information,
but of course no metadata tags.