Ampache: Support for the action 'tag'
This action to retrieve a single tag/genre was missed when adding the
the tags support in general. No known client uses it, but added for
completeness.

Also, fixed a slight bug in the formatting of the result of the `tags`
action: The `total_count` was shown as attribute of the root node, while
per the documentation
(https://github.com/ampache/ampache/wiki/XML-methods), it is supposed
to be shown as a child element.