Ampache: Add element `total_count` to all responses
As per the API documentation
https://github.com/ampache/ampache/wiki/XML-methods, all Ampache
responses describing songs, albums, artists, playlists, or tags, should
contain the `total_count` element. This has probably been added in the
API version 380001 (or maybe 400001?). We still advertise to support
only the API version 350001, as we probably have nowhere near a
complete support for 380001. However, it's a bit difficult to tell,
in which version has each minor improvement been introduced.