Fix unhandled exception in case the user session has closed
In case there is no logged in user, the ApiController constructor may
get called with $userFolder value null. This used to cause an
exception since the null didn't match the type hint "Folder".

The cloud core has a logic to redirect the user to the login page if
service requiring login is queried without a valid session. However,
this type mismatch exception apparently fired before that logic had a
chance to kick in.

This problem was probably also behind the report #682.