Fix 6th track showing twice during search on albums with 6 tracks
In case an album had exactly 6 tracks, and the 6th track matched search
in the Albums view, that track was being rendered twice.

This same problem probably affected also the Folders and Genres views,
in case the category in question had exactly 10 tracks, and the search
matched the 10th track.