Fix navigation from Files search results to Music
This seems to have got broken with commit 4157d65ce already on v0.13.0
(where the search was introduced). Initializing the maincontroller
`artists` variable to [] instead of null had the unforeseen side effect
of triggering the handling of URL parameters too early, before the
collection had been loaded.

The problem was present also in other cases where the user tried to
navigate to a Music app url containing the ID of track/album/artist.