Ampache/Subsonic: Fix small issues on the playlist track numbers
When listing the tracks of a playlist, the number of each track means the
ordinal of the track on the list in question, rather than its number on the
album. However, this logic had two bugs:

1. When the track belonged to a multi-disc album, then the number was
adjusted by adding (100 * disc #) to it, which made no sense in this context.

2. When fetching playlist tracks with an offset (possible via Ampache), the
numbers in the returned tracks didn't take the offset into account, i.e. the
numbering started from 1 on each page.