Subsonic: Add `created` and `comment` attributes to playlists
The `created` attribute is actually required by the Subsonic API
specification, so this was needed to be a conforming server.

The `created` and `comment` were also added as new fields to the
oc_music_playlists table. A migration script initializes the `created`
field with the installation date, so each playlist should always have some
valid data in the column. For now, these fields can be viewed only via the
Subsonic API, but they could be later added also to the web UI.