Subsonic: Clean up timestamp tags from synced lyrics, if those are used for `getLyrics`
The Subsonic service `getLyrics` primarily uses the unsynced lyrics if those
are available in the file tags. If only synced lyrics are available, then
those are parsed and converted into unsynced lyrics so that no timestamps or
other LRC tags are shown to the user. In case there are many timestamps on one
line, the line will be duplicated that many times and the lines are ordered
according the timestamps, for linear chronological representation.