Parse and strip the synced lyrics also when showing them in the details pane
In case the file in question has no unsynced lyrics in its metadata and it
does have time-synced lyrics, the details pane on the web UI now shows a
parsed and cleaned-up version of those synced lyrics. That is, the LRC tags
are not shown and the lines are ordered by the timestamps, duplicating lines
when necessary.

In case the file has time-synced lyrics available, those are now returned in
the response of `api/file/{id}/details` alongside the unsynced lyrics. For
now, they are not used for anything on the web UI, but in the future, the UI
could use this to show the lyrics so that the current line would be
highlighted.