Fix showing time-synced lyrics, when a timestamp is less than the offset
There was a problem in showing the time-synced lyrics in case one of
the timestamps was so small, that it became negative after the offset
adjustment. Such lyrics were laid out on the UI in an unexpected order
where all the rows with non-negative timestamps came first, and then
the negative timestamps.

Fixed by returning the lyrics from the backend in an ordered array
instead of an associative array. The associative array in practice
became a JSON object, where the iteration order in javascript is not
guaranteed. Also, all negative timestamps are now saturated to 0. This
wouldn't have been possible in the associative array, where timestamps
had to be unique keys.