Proper handling for the keyboard play/pause button on Firefox
The Firefox browser does not (yet) offer the mediaSession API for full
support of the multimedia keys on the keyboard. However, the newest
version 77 has support for the play/pause key without using the API.
Instead, the browser directly controls the playback state of the HTML5
audio element.

The PlayerWrapper class and the modules using it were now refactored
so, that they are prepared for external play/pause events on the audio
element. The updates of the UI playback state happen based on callbacks
from the audio element instead of directly updating the UI state after
issuing a command to the audio element.

The PlayerWrapper was refactored also so that the "toggle playback"
functionality is no longer its duty. The wrapper just offers play and
pause, and the toggling should happen by the higher level logic.