Fix an error in the browser console on NC caused by 16a2ed27
Nextcloud sets a Content-Security-Policy which bans the use of the
<base> tag. It seems that this can't even be overridden by an
application. Trying to use the <base> caused an error being printed to
the web browser console upon page load.

We were using the <base> tag only as a container to hold the base
address of the page. We were not actually relying on it doing anything,
so this caused no problems besides the error on console.

Fixed by not using the back-end utilities to generate the base address
for RESTAngular. Instead, the address is now evaluated using the
front-end utility method OC.generateUrl.