Use uploaded image as artist cover if the file name matches
In case the name of an image file uploaded to anywhere under the music
collection folder matches the name of an artist with no cover set, the newly
uploaded image is used as a cover image for the artist in question. The name
comparison is naturally made omitting the file extension. The matching is
case sensitive.

The file deleted hook now removes the artist cover image references, too,
when a cover image is deleted.