Use newly uploaded artist image even if the artist already has one set
The artist image used is now changed every time the user uploads an
image with a name exactly matching the artist name, even if the artist
in question already has an image set. Previously, it was rather
difficult to update the artist image.

Also, fixed several issues related to the artist image caching and
cache invalidation.