Fix Music's style definitions being applied to the Notifications popup
The definitions meant to be used with the "emptycontent" placeholder of
the Music app was being accidentally used also on the Notifications
popup provided by the Nextcloud server when it had no content.
Previously, this popup used to be hidden when there were no
notifications, but starting from ~Nextcloud 16, it has been shown with
the placeholder content.

There were two distinct styling issues caused by this:
1. The width of the Notifications popup was set to be so wide that
   most of its content fell off the screen (unless screen was very
   narrow)
2. The background color of the Notifications popup did not change
   correctly when the dark theme was activated.

refs #772