Use underlined instead of bolded style on the external links in Settings
There was such an issue, that the arrow symbol (↗) used on the links
could get rendered with an extremely strong highlight when bolded (it
got a blue box surrounding the arrow). This was a tad too flashy and
looked ugly especially in the orange warning box about Ampache/Subsonic
compatibility. Curiously, this was the case only on the combination of
Chrome and Nextcloud 17+. Using Firefox on NC17+ or Chrome on NC12-16,
the arrows were rendered "normally". So apparently, this has been
specifically changed by Nextcloud, but the style is not compatible with
Firefox.

Also the settings views of the Nextcloud and ownCloud cores use
underlined style for the links (even though the default in the core
css is the bolded style for links). Hence, this change also makes
the Music app more consistent with the rest of the cloud.