Support showing artist biography from Last.fm in the artist details
Opening the artist details view now shows the biography from Last.fm if
the `music.lastfm_api_key` has been set in the `config.php`. If the API
key is not set, a placeholder text is shown and its tooltip tells to
look into Settings view for further details.

The Settings view has a new section "Admin" which describes the options
available by editing `config.php`.

The library `angular-angular-sanitize` was added because it is needed
to use the directive `ng-bind-html`. This is used to show the biography.