Settings view: Add the arrow symbol to links usign css
Previously, the arrow symbol trailing all the external links was part
of the localizable content. Furthermore, it was missing from the
Last.fm link.

Now the arrow symbol is added to all exteranl links in Settings using
the style sheet. This is more appropriate since the arrow is more about
graphical style than logical content.