Ampache: Remove `total_count` from all results
This field has been deprecated in the Ampache XML API and is not
present at all in the specification of the JSON API. The real Ampache
server still keeps the field in the XML API until the API version 5.
However, the Music app isn't anyway fully compliant with the API
specification, and no known client uses this field. Hence, we drop the
field already now, to avoid differences between XML and JSON APIs.