Ampache: Improve the compliancy of the JSON API
The Ampache JSON API is still not frozen and the documentation still
has some internal inconsistencies. However, some details have got
clarified after our initial JSON API implementation. In this commit, I
have tried to match what is my best guess of the intention of the the
API documentation. Still, there are probably details which we still
need to refine in the future.

Changes made:
- All entity IDs are given using string type
- All other integer values are given using int type
- The root element in the JSON results is an anonymous array, exception
   being error responses where the root element is called `error`
- The values given in the XML API as text content of an element are in
   most cases mapped to JSON key called `name`. The error response is
   again an exception, and there, the key is called `message`.

refs #768