Refactor GenreMapper (affecting also BaseMapper and TrackMapper)
The GenreMapper and GenreBusinessLayer no longer have functions named
findAllWithCounts. Instead, all the standard getter functions from
BaseMapper like find and findAll now return the Genre entities with the
track/album/artist counts set. In addition to being more clean, this
fixes the behavior of the Ampache action `tag` which used to return the
requested tag without the related counts.

To make this work, we had to rule, that any override of
BaseMapper::selectEntities must not set alias name for the primary
table of the mapper. For this reason, also TrackMapper had to be
modified.